+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 * Add `Video` component. [#176](https://github.com/mapbox/dr-ui/pull/176)
+* Add `Browser` component. [#177](https://github.com/mapbox/dr-ui/pull/177)
 
 ## 0.20.1